# use javax.json streaming