Use the streaming version of the javax.json api to be comparable with the other streaming implementations.