@fabienrenaud 2020-12-25
1k serialization bench
```
> ./run ser --number 1 --size 1 --libs javaxjson,jackson --apis stream

Benchmark                                            Mode  Cnt        Score       Error  Units
c.g.f.j.stream.Serialization.jackson                thrpt   20  1313636.065 ± 94808.818  ops/s
c.g.f.j.stream.Serialization.javaxjson              thrpt   20    42442.679 ±  1822.025  ops/s
```

10kb deserialization bench
```
> ./run deser --number 1 --size 10 --libs javaxjson,jackson --apis stream
Benchmark                   Mode  Cnt       Score      Error  Units
Deserialization.jackson    thrpt   20  159284.778 ± 5469.069  ops/s
Deserialization.javaxjson  thrpt   20   33865.047 ±  529.688  ops/s
```

Seems to be about the same perf as before.

Merging.
Thank you for your contribution.