@tbutter 2020-12-25
@fabienrenaud I was a 10x increase on my computer. (./run ser --libs javaxjson,jackson --apis stream did go from about 4k to 43k ops/s)