@fabienrenaud 2020-12-26
@tbutter Can you post your results here before/after the change (with a benchmark of some other lib – like jackson – as a baseline for comp)?