@tbutter 2020-12-26
Before:

Benchmark                                            Mode  Cnt        Score       Error  Units
c.g.f.j.databind.Serialization.jackson              thrpt   20  1128372,325 ± 68940,272  ops/s
c.g.f.j.databind.Serialization.jackson_afterburner  thrpt   20  1145477,484 ± 89842,216  ops/s
c.g.f.j.stream.Serialization.jackson                thrpt   20  1183593,001 ± 71676,136  ops/s
c.g.f.j.stream.Serialization.javaxjson              thrpt   20     4884,346 ±   403,721  ops/s

After:

Result "com.github.fabienrenaud.jjb.stream.Serialization.javaxjson":
  50201,844 ±(99.9%) 1661,622 ops/s [Average]
  (min, avg, max) = (43844,098, 50201,844, 52512,946), stdev = 1913,526
  CI (99.9%): [48540,222, 51863,466] (assumes normal distribution)


# Run complete. Total time: 00:02:52


Benchmark                 Mode  Cnt      Score      Error  Units
Serialization.javaxjson  thrpt   20  50201,844 ± 1661,622  ops/s